% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection.R
\name{get.vim}
\alias{get.vim}
\title{Get variable importance.}
\usage{
get.vim(rf)
}
\arguments{
\item{rf}{Object of class \code{\link[ranger]{ranger}}}
}
\value{
numeric vector with importance value for each variable (in original order)
}
\description{
Extracts variable importance depending on class of random forest object.
}
