% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection_vita.R
\name{holdout.rf}
\alias{holdout.rf}
\title{Helper function for variable selection using Vita approach.}
\usage{
holdout.rf(
  x,
  y,
  ntree = 500,
  mtry.prop = 0.2,
  nodesize.prop = 0.1,
  no.threads = 1,
  type = "regression",
  importance = importance
)
}
\arguments{
\item{x}{matrix or data.frame of predictor variables with variables in
columns and samples in rows (Note: missing values are not allowed).}

\item{y}{vector with values of phenotype variable (Note: will be converted to factor if
classification mode is used).}

\item{ntree}{number of trees.}

\item{mtry.prop}{proportion of variables that should be used at each split.}

\item{nodesize.prop}{proportion of minimal number of samples in terminal
nodes.}

\item{no.threads}{number of threads used for parallel execution.}

\item{type}{mode of prediction ("regression", "classification" or "probability").}

\item{importance}{See \code{\link[ranger]{ranger}} for details.}
}
\value{
Hold-out random forests with variable importance
}
\description{
This function calculates a modified version of the permutation importance using two cross-validation folds (holdout folds)
as described in Janitza et al. (2015). Note that this function is a reimplementation of the \code{holdoutRF} function in the
R package \code{\link[ranger]{ranger}}.
}
\references{
Janitza, S., Celik, E. & Boulesteix, A.-L., (2015). A computationally fast variable importance test for random forest for high dimensional data, Technical Report 185, University of Munich, https://epub.ub.uni-muenchen.de/25587.
}
