% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{hybrid}
\alias{hybrid}
\alias{hybrid.default}
\alias{hybrid.formula}
\title{Feature selection with the hybrid algorithm}
\usage{
hybrid(x, ...)

\method{hybrid}{default}(
  x,
  y,
  pValue = 0.01,
  mcAdj = TRUE,
  maxRuns = 100,
  doTrace = 0,
  holdHistory = TRUE,
  getImp,
  alpha = 0.05,
  seed,
  ...
)

\method{hybrid}{formula}(formula, data = .GlobalEnv, ...)
}
\arguments{
\item{x}{data frame of predictors.}

\item{...}{additional parameters passed to \code{getImp}.}

\item{y}{response vector; factor for classification, numeric vector for regression, \code{Surv} object for survival (supports depends on importance adapter capabilities).}

\item{pValue}{confidence level. Default value should be used.}

\item{mcAdj}{if set to \code{TRUE}, a multiple comparisons adjustment using the Bonferroni method will be applied. Default value should be used; older (1.x and 2.x) versions of hybrid were effectively using \code{FALSE}.}

\item{maxRuns}{maximal number of importance source runs.
You may increase it to resolve attributes left Tentative.}

\item{doTrace}{verbosity level. 0 means no tracing, 1 means reporting decision about each attribute as soon as it is justified, 2 means the same as 1, plus reporting each importance source run, 3 means the same as 2, plus reporting of hits assigned to yet undecided attributes.}

\item{holdHistory}{if set to \code{TRUE}, the full history of importance is stored and returned as the \code{ImpHistory} element of the result.
Can be used to decrease a memory footprint of hybrid in case this side data is not used, especially when the number of attributes is huge; yet it disables plotting of such made \code{hybrid} objects and the use of the \code{\link{TentativeRoughFix}} function.}

\item{getImp}{function used to obtain attribute importance.
The default is get_imp_ranger, which runs random forest from the \code{ranger} package and gathers Z-scores of corrected impurities.
It should return a numeric vector of a size identical to the number of columns of its first argument, containing importance measure of respective attributes.
Any order-preserving transformation of this measure will yield the same result.
It is assumed that more important attributes get higher importance. +-Inf are accepted, NaNs and NAs are treated as 0s, with a warning.}

\item{alpha}{significance threshold used by Vita}

\item{seed}{Seed}

\item{formula}{alternatively, formula describing model to be analysed.}

\item{data}{in which to interpret formula.}
}
\value{
An object of class \code{hybrid}, which is a list with the following components:
\item{finalDecision}{a factor of three value: \code{Confirmed}, \code{Rejected} or \code{Tentative}, containing final result of feature selection.}
\item{ImpHistory}{a data frame of importances of attributes gathered in each importance source run.
Beside predictors' importances, it contains maximal, mean and minimal importance of shadow attributes in each run.
Rejected attributes get \code{-Inf} importance.
Set to \code{NULL} if \code{holdHistory} was given \code{FALSE}.}
\item{timeTaken}{time taken by the computation.}
\item{impSource}{string describing the source of importance, equal to a comment attribute of the \code{getImp} argument.}
\item{call}{the original call of the \code{hybrid} function.}
}
\description{
hybrid is an all relevant random forest feature selection wrapper algorithm that uses the corrected impurity importance (Nembrini et al. 2019) as variable importance measure (VIM);
Analogously to to Janitza et al. 2018, variables negative impurity importance are interpreted unimportant and used to generate the null distribution and to calculate the p-values.
That is, this implementation is fully based on the original implementation of the Vita variable selection algorithm available in \code{\link[ranger]{ranger}} and that of \code{\link[Boruta]{Boruta}}.
The method performs a top-down search for relevant features by comparing original attributes' importance with importance achievable at random, estimated using their permuted copies, and progressively eliminating irrelevant features to stabilise that test.
}
\details{
hybrid iteratively compares importances of attributes with importances of shadow attributes, created by shuffling original ones.
Attributes that have significantly worst importance than shadow ones are being consecutively dropped.
On the other hand, attributes that are significantly better than shadows are admitted to be Confirmed.
Shadows are re-created in each iteration.
Algorithm stops when only Confirmed attributes are left, or when it reaches \code{maxRuns} importance source runs.
If the second scenario occurs, some attributes may be left without a decision.
They are claimed Tentative.
You may try to extend \code{maxRuns} or lower \code{pValue} to clarify them, but in some cases their importances do fluctuate too much for hybrid to converge.
Instead, you can use \code{\link{TentativeRoughFix}} function, which will perform other, weaker test to make a final decision, or simply treat them as undecided in further analysis.
}
\examples{
set.seed(777)

\dontrun{
#hybrid on the "small redundant XOR" problem; read ?srx for details
data(srx)
hybrid(Y~.,data=srx)->hybrid.srx

#Results summary
print(hybrid.srx)

#Result plot
plot(hybrid.srx)

#Attribute statistics
attStats(hybrid.srx)

#Using alternative importance source, rFerns
hybrid(Y~.,data=srx,getImp=getImpFerns)->hybrid.srx.ferns
print(hybrid.srx.ferns)

#Verbose
hybrid(Y~.,data=srx,doTrace=2)->hybrid.srx
}
\dontrun{
#hybrid on the iris problem extended with artificial irrelevant features
#Generate said features
iris.extended<-data.frame(iris,apply(iris[,-5],2,sample))
names(iris.extended)[6:9]<-paste("Nonsense",1:4,sep="")
#Run hybrid on this data
hybrid(Species~.,data=iris.extended,doTrace=2)->hybrid.iris.extended
#Nonsense attributes should be rejected
print(hybrid.iris.extended)
}

\dontrun{
#hybrid on the HouseVotes84 data from mlbench
library(mlbench); data(HouseVotes84)
na.omit(HouseVotes84)->hvo
#Takes some time, so be patient
hybrid(Class~.,data=hvo,doTrace=2)->Bor.hvo
print(Bor.hvo)
plot(Bor.hvo)
plotImpHistory(Bor.hvo)
}
\dontrun{
#hybrid on the Ozone data from mlbench
library(mlbench); data(Ozone)
library(randomForest)
na.omit(Ozone)->ozo
hybrid(V4~.,data=ozo,doTrace=2)->Bor.ozo
cat('Random forest run on all attributes:\n')
print(randomForest(V4~.,data=ozo))
cat('Random forest run only on confirmed attributes:\n')
print(randomForest(ozo[,getSelectedAttributes(Bor.ozo)],ozo$V4))
}
\dontrun{
#hybrid on the Sonar data from mlbench
library(mlbench); data(Sonar)
#Takes some time, so be patient
hybrid(Class~.,data=Sonar,doTrace=2)->Bor.son
print(Bor.son)
#Shows important bands
plot(Bor.son,sort=FALSE)
}
}
\references{
Nembrini, S., Koenig, I. R. & Wright, M. N. (2018). The revival of the Gini Importance? Bioinformatics. https://doi.org/10.1093/bioinformatics/bty373.
Janitza, S, Celik, E, Boulesteix, AL. (2018). A computationally fast variable importance test for random forests for high-dimensional data. Adv Data Anal Classif.; doi.org: 10.1007/s11634-016-0276-4
Kursa, M. B. and Rudnicki, W. R. (2010). Feature Selection with the Boruta Package. Journal of Statistical Software. \emph{Journal of Statistical Software, 36(11)}, p. 1-13. URL: \url{http://www.jstatsoft.org/v36/i11/}.
}
