% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid.R
\name{print.hybrid}
\alias{print.hybrid}
\title{Print hybrid object}
\usage{
\method{print}{hybrid}(x, ...)
}
\arguments{
\item{x}{an object of a class hybrid.}

\item{...}{additional arguments passed to \code{\link{print}}.}
}
\value{
Invisible copy of \code{x}.
}
\description{
Print method for the hybrid objects.
}
