% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation.data.cor}
\alias{simulation.data.cor}
\title{Simulate gene expression data.}
\usage{
simulation.data.cor(no.samples, group.size, no.var.total, null = FALSE)
}
\arguments{
\item{no.samples}{number of samples.}

\item{group.size}{number of variables in each of the six groups of correlated 
variables.}

\item{no.var.total}{total number of variables.}

\item{null}{simulate null model (using independent functional variables).}
}
\value{
A data.frame with samples in rows and variables in columns (Note: first
  column contains simulated phenotype). Variables are named as y (= phenotype),
  g.i.j (= variable j in group i) and ind.k (= k-th independent variable).
}
\description{
Generates simulated gene expression data in a regression
setting.
}
\details{
The underlying simulation model is described in detail in the paper XXX. In
brief, a nonlinear regression model based on three uniformly distributed
variables is used. Predictor variables are simulated to be correlated with one
of those functional variables. In addition, independent, uniformly distributed
predictor variables are simulated.
}
\examples{
# simulate toy data set
data = simulation.data.cor(no.samples = 100, group.size = rep(10, 6), no.var.total = 200)
                     
}
