% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection_hybrid.R
\name{var.sel.hybrid}
\alias{var.sel.hybrid}
\title{Variable selection using hybrid function.}
\usage{
var.sel.hybrid(
  x,
  y,
  pValue = 0.01,
  maxRuns = 100,
  alpha = 0.05,
  ntree = 500,
  mtry.prop = 0.2,
  nodesize.prop = 0.1,
  no.threads = 1,
  method = "ranger",
  type = "regression",
  importance = "impurity_corrected",
  seed = 123,
  case.weights = NULL
)
}
\arguments{
\item{x}{matrix or data.frame of predictor variables with variables in
columns and samples in rows (Note: missing values are not allowed).}

\item{y}{vector with values of phenotype variable (Note: will be converted to factor if
classification mode is used).}

\item{pValue}{confidence level (default: 0.01 based on \code{\link[Boruta]{Boruta}} package)}

\item{maxRuns}{maximal number of importance source runs (default: 100 based on \code{\link[Boruta]{Boruta}} package)}

\item{alpha}{significance threshold used by Vita}

\item{ntree}{number of trees.}

\item{mtry.prop}{proportion of variables that should be used at each split.}

\item{nodesize.prop}{proportion of minimal number of samples in terminal
nodes.}

\item{no.threads}{number of threads used for parallel execution.}

\item{method}{implementation to be used ("ranger").}

\item{type}{mode of prediction ("regression", "classification" or "probability").}

\item{importance}{Variable importance mode ('none', 'impurity',
'impurity_corrected' or 'permutation'). Default is 'impurity_corrected'.}

\item{seed}{Seed}

\item{case.weights}{Weights for sampling of training observations. Observations with larger weights will be selected with higher probability in the bootstrap (or subsampled) samples for the trees.}
}
\value{
List with the following components:
  \itemize{
  \item \code{info} data.frame
  with information of each variable
  \itemize{
  \item run.x = original variable importance (VIM) in run x
  (includes min, mean and max of VIM of shadow variables)
  \item decision = Boruta decision (Confirmed, Rejected or Tentative)
  \item selected = variable has been selected
  }
  \item \code{var} vector of selected variables
  \item \code{info.shadow.var} data.frame with information about
  minimal, mean and maximal shadow variables of each run
  }

 @examples
# simulate toy data set
data = simulation.data.cor(no.samples = 100, group.size = rep(10, 6), no.var.total = 200)

# select variables
res = var.sel.hybrid(x = data[, -1], y = data[, 1])
res$var
}
\description{
Variable selection using the hybrid function and combining Vita and hybrid.
}
\details{
This function selects only variables that are confirmed based on hybrid implementation.
Note that this function uses the ranger implementation for variable selection.
}
