% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection_perm.R
\name{var.sel.perm}
\alias{var.sel.perm}
\title{Variable selection using a permutation approach.}
\usage{
var.sel.perm(
  x,
  y,
  no.perm = 100,
  p.t = 0,
  ntree = 500,
  mtry.prop = 0.2,
  nodesize.prop = 0.1,
  no.threads = 1,
  method = "ranger",
  type = "regression",
  parametric = FALSE,
  importance = "impurity_corrected",
  case.weights = NULL
)
}
\arguments{
\item{x}{matrix or data.frame of predictor variables with variables in
columns and samples in rows (Note: missing values are not allowed).}

\item{y}{vector with values of phenotype variable (Note: will be converted to factor if
classification mode is used).}

\item{no.perm}{number of permutations}

\item{p.t}{threshold for p-values (all variables with a p-value = 0  or < p.t will be selected)}

\item{ntree}{number of trees.}

\item{mtry.prop}{proportion of variables that should be used at each split.}

\item{nodesize.prop}{proportion of minimal number of samples in terminal
nodes.}

\item{no.threads}{number of threads used for parallel execution.}

\item{method}{implementation to be used ("ranger").}

\item{type}{mode of prediction ("regression", "classification" or "probability").}

\item{parametric}{logical stating if parametric permutation approach of Altmann et al. 2010 (based on normal
distribution) should be used (default: FALSE)}

\item{importance}{Variable importance mode ('none', 'impurity',
'impurity_corrected' or 'permutation'). Default is 'impurity_corrected'.}

\item{case.weights}{Weights for sampling of training observations. Observations with larger weights will be selected with higher probability in the bootstrap (or subsampled) samples for the trees.}
}
\value{
List with the following components:
  \itemize{
  \item \code{info} data.frame
  with information for each variable
  \itemize{
  \item vim.original = original variable importance (VIM)
  \item vim.perm.x = VIM in permutation x
  \item pvalue = proportion of permutations with a larger VIM than original VIM (nonparametric) or probability of observing the original or a larger VIM,
  given the fitted null importance distribution based on normal distributions (parametric)
  \item selected = variable has been selected
  }
  \item \code{var} vector of selected variables
  }
}
\description{
Selects variables which importance scores are larger than scores calculated
after permuting the phenotype. Output is a p-value calculated as the
proportion of permutations with an equal or larger importance score.
}
\details{
Note:
This function is a reimplementation of the approach in the R package \code{rfPermute} and the
parametric permutation approach by Altmann et al. (2010).
}
\references{
Altmann, A., Tolosi, L., Sander, O. & Lengauer, T. (2010). Permutation importance: a corrected feature importance measure, Bioinformatics 26:1340-1347.

 @examples
# simulate toy data set
data = simulation.data.cor(no.samples = 100, group.size = rep(10, 6), no.var.total = 200)

# select variables based on nonparametric permutation approach
res = var.sel.perm(x = data[, -1], y = data[, 1], no.perm = 10, p.t = 0)
res$var

# select variables based on parametric permutation approach
res.par = var.sel.perm(x = data[, -1], y = data[, 1], no.perm = 10, p.t = 0.05, parametric = TRUE)
res.par$var
}
