% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection_r2vim.R
\name{var.sel.r2vim}
\alias{var.sel.r2vim}
\title{Variable selection using recurrent relative variable importance (r2VIM).}
\usage{
var.sel.r2vim(
  x,
  y,
  no.runs = 10,
  factor = 1,
  ntree = 500,
  mtry.prop = 0.2,
  nodesize.prop = 0.1,
  no.threads = 1,
  method = "ranger",
  type = "regression",
  importance = "impurity_corrected",
  case.weights = NULL
)
}
\arguments{
\item{x}{matrix or data.frame of predictor variables with variables in
columns and samples in rows (Note: missing values are not allowed).}

\item{y}{vector with values of phenotype variable (Note: will be converted to factor if
classification mode is used).}

\item{no.runs}{number of random forests to be generated}

\item{factor}{minimal relative importance score for a variable to be selected}

\item{ntree}{number of trees.}

\item{mtry.prop}{proportion of variables that should be used at each split.}

\item{nodesize.prop}{proportion of minimal number of samples in terminal
nodes.}

\item{no.threads}{number of threads used for parallel execution.}

\item{method}{implementation to be used ("ranger").}

\item{type}{mode of prediction ("regression", "classification" or "probability").}

\item{importance}{Variable importance mode ('none', 'impurity',
'impurity_corrected' or 'permutation'). Default is 'impurity_corrected'.}

\item{case.weights}{Weights for sampling of training observations. Observations with larger weights will be selected with higher probability in the bootstrap (or subsampled) samples for the trees.}
}
\value{
List with the following components:
  \itemize{
  \item \code{info} data.frame
  with information for each variable
  \itemize{
  \item vim.run.x = original variable importance (VIM) in run x
  \item rel.vim.run.x = relative VIM in run x
  \item rel.vim.min = minimal relative VIM over all runs
  \item rel.vim.med = median relative VIM over all runs
  \item selected = variable has been selected
  }
  \item \code{var} vector of selected variables
  }

 @examples
# simulate toy data set
data = simulation.data.cor(no.samples = 100, group.size = rep(10, 6), no.var.total = 200)

# select variables
res = var.sel.r2vim(x = data[, -1], y = data[, 1], no.runs = 5, factor = 1)
res$var
}
\description{
Generates several random forests using all variables and different random
number seeds. For each run, the importance score is divided by the (absolute)
minimal importance score (relative importance scores). Variables are selected
if the minimal relative importance score is >= factor.
}
\details{
Note: This function is a reimplementation of the R package \code{RFVarSelGWAS}.
}
