% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection_rfe.R
\name{var.sel.rfe}
\alias{var.sel.rfe}
\title{Variable selection using recursive feature elimination.}
\usage{
var.sel.rfe(
  x,
  y,
  prop.rm = 0.2,
  recalculate = TRUE,
  tol = 10,
  ntree = 500,
  mtry.prop = 0.2,
  nodesize.prop = 0.1,
  no.threads = 1,
  method = "ranger",
  type = "regression",
  importance = "impurity_corrected",
  case.weights = NULL
)
}
\arguments{
\item{x}{matrix or data.frame of predictor variables with variables in
columns and samples in rows (Note: missing values are not allowed).}

\item{y}{vector with values of phenotype variable (Note: will be converted to factor if
classification mode is used).}

\item{prop.rm}{proportion of variables removed at each step (default value of \code{\link[varSelRF]{varSelRF}})}

\item{recalculate}{logical stating if importance should be recalculated at each iteration (default: TRUE)}

\item{tol}{acceptable difference in optimal performance (finds the smallest subset size that has a percent loss less than tol)}

\item{ntree}{number of trees.}

\item{mtry.prop}{proportion of variables that should be used at each split.}

\item{nodesize.prop}{proportion of minimal number of samples in terminal
nodes.}

\item{no.threads}{number of threads used for parallel execution.}

\item{method}{implementation to be used ("ranger").}

\item{type}{mode of prediction ("regression", "classification" or "probability").}

\item{importance}{Variable importance mode ('none', 'impurity',
'impurity_corrected' or 'permutation'). Default is 'impurity_corrected'.}

\item{case.weights}{Weights for sampling of training observations. Observations with larger weights will be selected with higher probability in the bootstrap (or subsampled) samples for the trees.}
}
\value{
List with the following components:
  \itemize{
  \item \code{info} data.frame
  with information for each variable
  \itemize{
  \item included.until.subset = number of smallest subset which contains variable
  \item selected = variable has been selected
  }
  \item \code{var} vector of selected variables
  \item \code{info.runs} data.frame with information for each run
  \itemize{
  \item n = number of variables
  \item mse = mean squared error
  \item rsq = R^2
  }}

 @examples
# simulate toy data set
data = simulation.data.cor(no.samples = 100, group.size = rep(10, 6), no.var.total = 200)

# select variables
res = var.sel.rfe(x = data[, -1], y = data[, 1], prop.rm = 0.2, recalculate = TRUE)
res$var
}
\description{
Compares random forests based on nested subsets of the variables and selects
those variables leading to the forest with the smallest prediction error within a tolerance.
}
\details{
Note: This function differs from the approach implemented in the R package
\code{\link[varSelRF]{varSelRF}} because it recalculates importance scores in each step. The tolerance step is based on the
\code{pickSizeTolerance} function in the R package \code{caret}.
}
