% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection_vita.R
\name{var.sel.vita}
\alias{var.sel.vita}
\title{Variable selection using Vita approach.}
\usage{
var.sel.vita(
  x,
  y,
  p.t = 0.05,
  ntree = 500,
  mtry.prop = 0.2,
  nodesize.prop = 0.1,
  no.threads = 1,
  method = "ranger",
  type = "regression",
  importance = "impurity_corrected"
)
}
\arguments{
\item{x}{matrix or data.frame of predictor variables with variables in
columns and samples in rows (Note: missing values are not allowed).}

\item{y}{vector with values of phenotype variable (Note: will be converted to factor if
classification mode is used).}

\item{p.t}{threshold for p-values (all variables with a p-value = 0  or < p.t will be selected)}

\item{ntree}{number of trees.}

\item{mtry.prop}{proportion of variables that should be used at each split.}

\item{nodesize.prop}{proportion of minimal number of samples in terminal
nodes.}

\item{no.threads}{number of threads used for parallel execution.}

\item{method}{implementation to be used ("ranger").}

\item{type}{mode of prediction ("regression", "classification" or "probability").}

\item{importance}{Variable importance mode ('none', 'impurity',
'impurity_corrected' or 'permutation'). Default is 'impurity_corrected'.}
}
\value{
List with the following components:
  \itemize{
  \item \code{info} data.frame
  with information for each variable
  \itemize{
  \item vim = variable importance (VIM)
  \item CI_lower = lower confidence interval boundary
  \item CI_upper = upper confidence interval boundary
  \item pvalue = empirical p-value
  \item selected = variable has been selected
  }
  \item \code{var} vector of selected variables
  }
 @references
 Janitza, S., Celik, E. & Boulesteix, A.-L., (2015). A computationally fast variable importance test for random forest for high dimensional data, Technical Report 185, University of Munich, https://epub.ub.uni-muenchen.de/25587.

 @examples
# simulate toy data set
data = simulation.data.cor(no.samples = 100, group.size = rep(10, 6), no.var.total = 500)

# select variables
res = var.sel.vita(x = data[, -1], y = data[, 1], p.t = 0.05)
res$var
}
\description{
This function calculates p-values based on the empirical null distribution from non-positive VIMs as
described in Janitza et al. (2015). Note that this function uses the \code{importance_pvalues} function in the R package
\code{\link[ranger]{ranger}}.
}
