% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection.R
\name{wrapper.rf}
\alias{wrapper.rf}
\title{Wrapper function to call random forests function.}
\usage{
wrapper.rf(x, y, ntree = 500, mtry.prop = 0.2, nodesize.prop = 0.1,
  no.threads = 1, method = "ranger", type = "regression", ...)
}
\arguments{
\item{x}{matrix or data.frame of predictor variables with variables in
columns and samples in rows (Note: missing values are not allowed).}

\item{y}{vector with values of phenotype variable (Note: will be converted to factor if
classification mode is used).}

\item{ntree}{number of trees.}

\item{mtry.prop}{proportion of variables that should be used at each split.}

\item{nodesize.prop}{proportion of minimal number of samples in terminal
nodes.}

\item{no.threads}{number of threads used for parallel execution.}

\item{method}{implementation to be used ("ranger").}

\item{type}{mode of prediction ("regression", "classification" or "probability").}

\item{...}{further arguments needed for \code{\link[relVarId]{holdout.rf}} function only.}
}
\value{
An object of class \code{\link[ranger]{ranger}}.
}
\description{
Provides an interface to different parallel implementations of the random
forest algorithm. Currently, only the \code{ranger} package is
supported.
}
\examples{
# simulate toy data set
data = simulation.data.cor(no.samples = 100, group.size = rep(10, 6), no.var.total = 200)

# regression
wrapper.rf(x = data[, -1], y = data[, 1],
           type = "regression", method = "ranger")
}

